server/requirements.txt
flask
pillow
openai>=1.0.0
